:sparkles: #152 can now switch between groups
