+++ frontend/src/stores/punishmentToFilter.ts
@@ -3,7 +3,9 @@ import type { PunishmentType } from '../lib/types'
 
 function getRawPunishmentFilters(): string {
   const rawPunishmentFilters = window.localStorage.getItem('punishmentFilters')
-  return rawPunishmentFilters !== 'null' ? rawPunishmentFilters : '[]'
+  return rawPunishmentFilters === 'null' || rawPunishmentFilters == 'undefined'
+    ? '[]'
+    : rawPunishmentFilters
 }
 
 export const punishmentsToFilter = writable<PunishmentType[]>(
